fixed failing tests in pandas 0.21.0.rc1
* Probably due to https://pandas-docs.github.io/pandas-docs-travis/whatsnew.html#dtype-conversions
* credit to Liesbeth for finding

+++ brainbuilder/brainbuilder/utils/bbp.py
@@ -475,7 +475,7 @@ def get_placement_hints_table(morphs):
         scores = np.array(list(itertools.chain(*hints_group.values))).reshape((count, length))
 
         # placement hints are organised bottom (high score) to top (low score)
-        scores = np.fliplr(scores)
+        scores = np.fliplr(scores).astype(np.float)
 
         repetitions = [subdivision_count // length] * length
         extended = np.repeat(scores, repetitions, axis=1)
